+++ src/Smartstore.Core/Catalog/Search/Configuration/SearchSettings.cs
@@ -1,6 +1,7 @@
 ﻿using Smartstore.Core.Catalog.Products;
 using Smartstore.Core.Configuration;
 using Smartstore.Core.Search;
+using Smartstore.Core.Search.Facets;
 
 namespace Smartstore.Core.Catalog.Search
 {
@@ -61,8 +62,6 @@ namespace Smartstore.Core.Catalog.Search
         /// </summary>
         public bool SearchProductByIdentificationNumber { get; set; }
 
-        // TBD: what about area specific searchin setting (product, blog, etc.)
-
         #region Common facet settings
 
         /// <summary>
@@ -83,6 +82,10 @@ namespace Smartstore.Core.Catalog.Search
         public int DeliveryTimeDisplayOrder { get; set; } = 4;
         public int AvailabilityDisplayOrder { get; set; } = 5;
         public int NewArrivalsDisplayOrder { get; set; } = 6;
+
+        public FacetSorting CategorySorting { get; set; } = FacetSorting.HitsDesc;
+        public FacetSorting BrandSorting { get; set; } = FacetSorting.LabelAsc;
+        public FacetSorting DeliveryTimeSorting { get; set; } = FacetSorting.DisplayOrder;
 
         #endregion