+++ src/Smartstore.Core/Migrations/SmartDbContextDataSeeder.cs
@@ -10,18 +10,24 @@ namespace Smartstore.Core.Data.Migrations
 
         public async Task SeedAsync(SmartDbContext context, CancellationToken cancelToken = default)
         {
-            //await context.MigrateLocaleResourcesAsync(MigrateLocaleResources);
+            await context.MigrateLocaleResourcesAsync(MigrateLocaleResources);
             //await MigrateSettingsAsync(context, cancelToken);
         }
 
-        public async Task MigrateSettingsAsync(SmartDbContext db, CancellationToken cancelToken = default)
-        {
+        //public async Task MigrateSettingsAsync(SmartDbContext db, CancellationToken cancelToken = default)
+        //{
 
-        }
+        //}
 
         public void MigrateLocaleResources(LocaleResourcesBuilder builder)
         {
-            
+            builder.AddOrUpdate("Admin.Configuration.Settings.Search.CommonFacet.Sorting",
+                "Sorting",
+                "Sortierung",
+                "Specifies the sorting of the search filters.",
+                "Legt die Sortierung der Suchfilter fest.");
+
+            builder.AddOrUpdate("Enums.FacetSorting.ValueAsc", "Value/ID: lowest first", "Wert/ID: Niedrigste zuerst");
         }
     }
 }